compilation warning fix

+++ libEM/pointarray.cpp
@@ -1072,17 +1072,17 @@ EMData *PointArray::pdb2mrc_by_summation(int map_size, float apix, float res)
 			kmax = map_size;
 
 		for (int k = kmin; k < kmax; k++) {
-			int table_index_z = int (fabs(k - zc) * inv_table_step_size);
+			size_t table_index_z = size_t (fabs(k - zc) * inv_table_step_size);
 			if ( table_index_z >= table.size() ) continue;
 			double zval = table[table_index_z];
 			size_t pd_index_z = k * map_size * map_size;
 			for (int j = jmin; j < jmax; j++) {
-				int table_index_y = int (fabs(j - yc) * inv_table_step_size);
+				size_t table_index_y = size_t (fabs(j - yc) * inv_table_step_size);
 				if ( table_index_y >= table.size() ) continue;
 				double yval = table[table_index_y];
 				size_t pd_index = pd_index_z + j * map_size + imin;
 				for (int i = imin; i < imax; i++, pd_index++) {
-					int table_index_x = int (fabs(i - xc) * inv_table_step_size);
+					size_t table_index_x = size_t (fabs(i - xc) * inv_table_step_size);
 					if ( table_index_x >= table.size() ) continue;
 					double xval = table[table_index_x];
 					pd[pd_index] += (float) (fval * zval * yval * xval);
